Move GeocodeResult class back to geo module
